+++ wireshell
@@ -10,13 +10,11 @@ use Wireshell\Commands\CreateTemplateCommand;
 use Wireshell\Commands\ServeCommand;
 use Wireshell\Commands\ShowVersionCommand;
 
-use GuzzleHttp\Client;
-
 require 'vendor/autoload.php';
 
 $app = new Application('WireShell - An extendable ProcessWire CLI', '0.2.1');
 
-$app->add(new NewCommand(new Client));
+$app->add(new NewCommand());
 $app->add(new CreateUserCommand());
 $app->add(new CreateRoleCommand());
 $app->add(new CreateTemplateCommand());
